feat: 34 ~ 36 강 commit  - 34강. 다양한 SQL을 알아보자!  - 35강. 애플리케이션 대신 DB로 기능 구현하기  - 36강. 세 번째 요구사항 클리어!

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -4,7 +4,7 @@ import com.group.libraryapp.domain.book.BookType
 
 data class BookStatResponse (
     val type: BookType,
-    var count: Int,
+    var count: Long,
 ) {
     fun plusOne() {
         count++
